# api/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "오픈마켓 API",
    "description": "오픈마켓 API Server입니다.<br><a href=\"/\">버전별 변경사항 확인</a>"
  },
  "servers": [
    {
      "url": "https://market-lion.koyeb.app/api",
      "description": ""
    },
    {
      "url": "http://localhost/api",
      "description": "로컬 테스트"
    }
  ],
...
  "paths": {
    "/users/": {
      "post": {
        "tags": [
          "회원"
        ],
        "summary": "회원 가입",
        "description": "회원 가입을 합니다.<br>회원 가입을 완료한 후 회원 정보를 반환합니다.",
        "responses": {
          "201": {
            "description": "성공",
            "content": {
              "application/json": {
                "examples": {
                  "기본 속성": {
                    "$ref": "#/components/examples/createUserRes"
                  },
                  "extra 속성": {
                    "$ref": "#/components/examples/createUserResWithExtra"
                  }
                }
              }
            }
          },
          "409": {
            "description": "이메일 중복됨",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error409"
                }
              }
            }
          },
          "422": {
            "description": "입력값 검증 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error422"
                }
              }
            }
          },
          "500": {
            "description": "서버 에러",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error500"
                }
              }
            }
          }
        },
        "requestBody": {
          "description": "회원 정보",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#components/schemas/createUser"
              },
              "examples": {
                "기본 속성": {
                  "$ref": "#/components/examples/createUser"
                },
                "extra 속성": {
                  "$ref": "#/components/examples/createUserWithExtra"
                }
              }
            }
          }
        }
      }
    },
...
    "/users/{_id}/{*}": {
      "get": {
        "tags": [
          "회원"
        ],
        "summary": "회원 정보 조회(단일 속성)",
        "description": "회원 정보 중 한가지 속성을 조회한다.",
        "parameters": [
          {
            "name": "_id",
            "description": "조회할 회원 id",
            "in": "path",
            "example": "5",
            "required": true,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "*",
            "description": "조회할 속성<br>단일 속성 조회 예시: name<br>중첩 속성 조회 예시: extra/addressBook",
            "in": "path",
            "example": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "examples": {
                  "users/5/name 조회": {
                    "$ref": "#/components/examples/userInfoResOneAttr"
                  },
                  "users/5/extra/addressBook 조회": {
                    "$ref": "#/components/examples/userInfoResWithExtra"
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error401"
                }
              }
            }
          },
          "404": {
            "description": "회원이 존재하지 않거나 접근 권한 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error404"
                }
              }
            }
          },
          "500": {
            "description": "서버 에러",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error500"
                }
              }
            }
          }
        },
        "security": [
          {
            "Access Token": []
          }
        ]
      }
    },
    "/users/{_id}": {
      "get": {
        "tags": [
          "회원"
        ],
        "summary": "회원 정보 조회(모든 속성)",
        "description": "회원 정보의 모든 속성을 조회한다.",
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "조회할 회원 id",
            "example": "5"
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "examples": {
                  "기본 속성": {
                    "$ref": "#/components/examples/createUserRes"
                  },
                  "extra 속성": {
                    "$ref": "#/components/examples/createUserResWithExtra"
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error401"
                }
              }
            }
          },
          "404": {
            "description": "회원이 존재하지 않거나 접근 권한 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error404"
                }
              }
            }
          },
          "500": {
            "description": "서버 에러",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error500"
                }
              }
            }
          }
        },
        "security": [
          {
            "Access Token": []
          }
        ]
      },
      "patch": {
        "tags": [
          "회원"
        ],
        "summary": "회원 정보 수정",
        "description": "회원 정보를 수정한다.",
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "회원 id",
            "example": "5"
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "examples": {
                  "기본 속성": {
                    "$ref": "#/components/examples/updateUserResOneAttr"
                  },
                  "extra 속성": {
                    "$ref": "#/components/examples/updateUserResWithExtra"
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error401"
                }
              }
            }
          },
          "404": {
            "description": "회원이 존재하지 않거나 접근 권한 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error404"
                }
              }
            }
          },
          "500": {
            "description": "서버 에러",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error500"
                }
              }
            }
          }
        },
        "security": [
          {
            "Access Token": []
          }
        ],
        "requestBody": {
          "description": "수정할 회원 정보",
          "required": true,
          "content": {
            "application/json": {
              "examples": {
                "기본 속성": {
                  "$ref": "#/components/examples/updateUserOneAttr"
                },
                "extra 속성": {
                  "$ref": "#/components/examples/updateUserWithExtra"
                }
              }
            }
          }
        }
      }
    },
...
    "/posts/users/{_id}": {
      "get": {
        "tags": [
          "게시판"
        ],
        "summary": "사용자 게시물 목록",
        "description": "사용자가 작성한 게시물 목록을 조회한다.",
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "Access Token": []
          }
        ]
      }
    },
...
    "/admin/users/": {
      "get": {
        "tags": [
          "회원 관리"
        ],
        "summary": "회원 목록 조회",
        "description": "회원 목록을 조회한다.",
        "parameters": [
          {
            "name": "_id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "email",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "phone",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "address",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "custom",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "Access Token": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "Access Token": {
        "type": "http",
        "scheme": "bearer",
        "in": "header",
        "bearerFormat": "JWT"
      },
      "Refresh Token": {
        "type": "http",
        "scheme": "bearer",
        "in": "header",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "error401": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "number",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "{인증 실패 사유}"
          },
          "errorName": {
            "type": "string",
            "example": "EmptyAuthorization | TokenExpiredError | JsonWebTokenError"
          }
        }
      },
...
      "error404": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "number",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "/api/xxx 리소스를 찾을 수 없습니다."
          }
        }
      },
      "error409": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "number",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "이미 등록된 리소스입니다."
          }
        }
      },
      "error422": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "number",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "잘못된 입력값이 있습니다."
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "field"
                },
                "value": {
                  "type": "string",
                  "example": "swaggermarket.com"
                },
                "msg": {
                  "type": "string",
                  "example": "이메일 형식에 맞지 않습니다."
                },
                "path": {
                  "type": "string",
                  "example": "email"
                },
                "location": {
                  "type": "string",
                  "example": "body"
                }
              }
            }
          }
        }
      },
      "error500": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "number",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "요청하신 작업 처리에 실패했습니다. 잠시 후 다시 이용해 주시기 바랍니다."
          }
        }
      },
...
    },
    "examples": {
...
      "createUser": {
        "value": {
          "email": "swagger@market.com",
          "password": "12345678",
          "name": "스웨거",
          "type": "user",
          "phone": "01011112222",
          "address": "서울시 강남구 역삼동 123"
        }
      },
      "createUserWithExtra": {
        "value": {
          "email": "swagger@market.com",
          "password": "12345678",
          "name": "스웨거",
          "type": "user",
          "phone": "01011112222",
          "address": "서울시 강남구 역삼동 123",
          "extra": {
            "gender": "extra에는 프로젝트에서 필요한 아무 속성이나",
            "age": "필요한 값을",
            "address": [
              "넣으면",
              "됩니다."
            ],
            "profileImage": "/uploads/swagger.jpg",
            "obj": {
              "hello": "객체로",
              "hi": "넣어도 됩니다."
            },
            "addressBook": [
              {
                "name": "집",
                "address": "서울시"
              },
              {
                "name": "회사",
                "address": "인천시"
              }
            ]
          }
        }
      },
      "createUserRes": {
        "value": {
          "ok": 1,
          "item": {
            "email": "swagger@market.com",
            "name": "스웨거",
            "type": "user",
            "phone": "01011112222",
            "address": "서울시 강남구 역삼동 123",
            "_id": 5,
            "createdAt": "2023.11.21 14:31:15",
            "updatedAt": "2023.11.21 14:31:15"
          }
        }
      },
      "createUserResWithExtra": {
        "value": {
          "ok": 1,
          "item": {
            "email": "swagger@market.com",
            "name": "스웨거",
            "type": "user",
            "phone": "01011112222",
            "address": "서울시 강남구 역삼동 123",
            "extra": {
              "gender": "extra에는 프로젝트에서 필요한 아무 속성이나",
              "age": "필요한 값을",
              "address": [
                "넣으면",
                "됩니다."
              ],
              "profileImage": "/uploads/swagger.jpg",
              "obj": {
                "hello": "객체로",
                "hi": "넣어도 됩니다."
              },
              "addressBook": [
                {
                  "name": "집",
                  "address": "서울시"
                },
                {
                  "name": "회사",
                  "address": "인천시"
                }
              ]
            },
            "_id": 5,
            "createdAt": "2023.11.21 14:33:41",
            "updatedAt": "2023.11.21 14:33:41"
          }
        }
      },
      "userInfoResOneAttr": {
        "value": {
          "ok": 1,
          "item": {
            "name": "스웨거"
          }
        }
      },
      "userInfoResWithExtra": {
        "value": {
          "ok": 1,
          "item": {
            "extra": {
              "addressBook": [
                {
                  "name": "집",
                  "address": "서울시"
                },
                {
                  "name": "회사",
                  "address": "인천시"
                }
              ]
            }
          }
        }
      },
      "updateUserOneAttr": {
        "value": {
          "phone": "01099998888",
          "name": "스웨거2"
        }
      },
      "updateUserResOneAttr": {
        "value": {
          "ok": 1,
          "updated": {
            "name": "스웨거2",
            "phone": "01099998888",
            "updatedAt": "2023.11.21 20:16:53"
          }
        }
      },
      "updateUserWithExtra": {
        "value": {
          "extra": {
            "address": [
              {
                "id": 1,
                "name": "회사",
                "value": "서울시 강남구 삼성동 111"
              },
              {
                "id": 2,
                "name": "학교",
                "value": "서울시 강남구 역삼동 222"
              }
            ]
          }
        }
      },
      "updateUserResWithExtra": {
        "value": {
          "ok": 1,
          "updated": {
            "extra": {
              "address": [
                {
                  "id": 1,
                  "name": "회사",
                  "value": "서울시 강남구 삼성동 111"
                },
                {
                  "id": 2,
                  "name": "학교",
                  "value": "서울시 강남구 역삼동 222"
                }
              ]
            },
            "updatedAt": "2023.11.21 20:13:33"
          }
        }
      }
    }
  }
}